Script Engine Update Phase 1
Dropping array limits, test phase.
Also Implements a under-development (u)int64 key support to the dbmaps.
Made Possible Thanks to Haruna!

+++ src/map/mapreg_sql.c
@@ -16,8 +16,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-static DBMap* mapreg_db = NULL; // int var_id -> int value
-static DBMap* mapregstr_db = NULL; // int var_id -> char* value
+static DBMap* mapreg_db = NULL; // int64 var_id -> int value
+static DBMap* mapregstr_db = NULL; // int64 var_id -> char* value
 static struct eri *mapreg_ers; //[Ind/Hercules]
 
 static char mapreg_table[32] = "mapreg";
@@ -26,30 +26,31 @@ static bool mapreg_str_dirty = false;
 
 #define MAPREG_AUTOSAVE_INTERVAL (300*1000)
 
+/* TODO may be worth not to save/store arrays with the SCRIPT_ARRAY_LEN_IDX index */
 
 /// Looks up the value of an integer variable using its uid.
-int mapreg_readreg(int uid) {
-	struct mapreg_save *m = idb_get(mapreg_db, uid);
+int mapreg_readreg(int64 uid) {
+	struct mapreg_save *m = i64db_get(mapreg_db, uid);
 	return m?m->u.i:0;
 }
 
 /// Looks up the value of a string variable using its uid.
-char* mapreg_readregstr(int uid) {
-	struct mapreg_save *m = idb_get(mapregstr_db, uid);
+char* mapreg_readregstr(int64 uid) {
+	struct mapreg_save *m = i64db_get(mapregstr_db, uid);
 	return m?m->u.str:NULL;
 }
 
 /// Modifies the value of an integer variable.
-bool mapreg_setreg(int uid, int val) {
+bool mapreg_setreg(int64 uid, int val) {
 	struct mapreg_save *m;
-	int num = (uid & 0x00ffffff);
-	int i   = (uid & 0xff000000) >> 24;
+	int num = (int32)(uid & 0xFFFFFFFF);
+	int i   = (int32)((uid >> 32) & 0xFFFFFFFF);
 	const char* name = script->get_str(num);
 
 	if( val != 0 ) {
-		if( (m = idb_get(mapreg_db,uid)) ) {
+		if( (m = i64db_get(mapreg_db,uid)) ) {
 			m->u.i = val;
-			if(name[1] != '@') {
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 				m->save = true;
 				mapreg_i_dirty = true;
 			}
@@ -60,21 +61,23 @@ bool mapreg_setreg(int uid, int val) {
 			m->uid = uid;
 			m->save = false;
 
-			if(name[1] != '@') {// write new variable to database
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
+				// write new variable to database
 				char tmp_str[32*2+1];
 				SQL->EscapeStringLen(mmysql_handle, tmp_str, name, strnlen(name, 32));
 				if( SQL_ERROR == SQL->Query(mmysql_handle, "INSERT INTO `%s`(`varname`,`index`,`value`) VALUES ('%s','%d','%d')", mapreg_table, tmp_str, i, val) )
 					Sql_ShowDebug(mmysql_handle);
 			}
-			idb_put(mapreg_db, uid, m);
+			i64db_put(mapreg_db, uid, m);
 		}
 	} else { // val == 0
-		if( (m = idb_get(mapreg_db,uid)) ) {
+		if( (m = i64db_get(mapreg_db,uid)) ) {
 			ers_free(mapreg_ers, m);
 		}
-		idb_remove(mapreg_db,uid);
+		i64db_remove(mapreg_db,uid);
 
-		if( name[1] != '@' ) {// Remove from database because it is unused.
+		if( name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
+			// Remove from database because it is unused.
 			if( SQL_ERROR == SQL->Query(mmysql_handle, "DELETE FROM `%s` WHERE `varname`='%s' AND `index`='%d'", mapreg_table, name, i) )
 				Sql_ShowDebug(mmysql_handle);
 		}
@@ -84,29 +87,29 @@ bool mapreg_setreg(int uid, int val) {
 }
 
 /// Modifies the value of a string variable.
-bool mapreg_setregstr(int uid, const char* str) {
+bool mapreg_setregstr(int64 uid, const char* str) {
 	struct mapreg_save *m;
-	int num = (uid & 0x00ffffff);
-	int i   = (uid & 0xff000000) >> 24;
+	int num = (int)(uid & 0xFFFFFFFF);
+	int i   = (int32)((uid >> 32) & 0xFFFFFFFF);
 	const char* name = script->get_str(num);
 	
 	if( str == NULL || *str == 0 ) {
-		if(name[1] != '@') {
+		if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 			if( SQL_ERROR == SQL->Query(mmysql_handle, "DELETE FROM `%s` WHERE `varname`='%s' AND `index`='%d'", mapreg_table, name, i) )
 				Sql_ShowDebug(mmysql_handle);
 		}
-		if( (m = idb_get(mapregstr_db,uid)) ) {
+		if( (m = i64db_get(mapregstr_db,uid)) ) {
 			if( m->u.str != NULL )
 				aFree(m->u.str);
 			ers_free(mapreg_ers, m);
 		}
-		idb_remove(mapregstr_db,uid);
+		i64db_remove(mapregstr_db,uid);
 	} else {
-		if( (m = idb_get(mapregstr_db,uid)) ) {
+		if( (m = i64db_get(mapregstr_db,uid)) ) {
 			if( m->u.str != NULL )
 				aFree(m->u.str);
 			m->u.str = aStrdup(str);
-			if(name[1] != '@') {
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 				mapreg_str_dirty = true;
 				m->save = true;
 			}
@@ -117,7 +120,8 @@ bool mapreg_setregstr(int uid, const char* str) {
 			m->u.str = aStrdup(str);
 			m->save = false;
 			
-			if(name[1] != '@') { //put returned null, so we must insert.
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
+				//put returned null, so we must insert.
 				char tmp_str[32*2+1];
 				char tmp_str2[255*2+1];
 				SQL->EscapeStringLen(mmysql_handle, tmp_str, name, strnlen(name, 32));
@@ -125,7 +129,7 @@ bool mapreg_setregstr(int uid, const char* str) {
 				if( SQL_ERROR == SQL->Query(mmysql_handle, "INSERT INTO `%s`(`varname`,`index`,`value`) VALUES ('%s','%d','%s')", mapreg_table, tmp_str, i, tmp_str2) )
 					Sql_ShowDebug(mmysql_handle);
 			}
-			idb_put(mapregstr_db, uid, m);
+			i64db_put(mapregstr_db, uid, m);
 		}
 	}
 
@@ -162,28 +166,60 @@ static void script_load_mapreg(void) {
 		struct mapreg_save *m = NULL;
 		int s = script->add_str(varname);
 		int i = index;
+		DBMap *mdb = mapreg_db;
 
 		if( varname[length-1] == '$' ) {
-			if( idb_exists(mapregstr_db, (i<<24)|s) ) {
-				ShowWarning("load_mapreg: duplicate! '%s' => '%s' skipping...\n",varname,value);
-				continue;
-			}
-		} else {
-			if( idb_exists(mapreg_db, (i<<24)|s) ) {
-				ShowWarning("load_mapreg: duplicate! '%s' => '%s' skipping...\n",varname,value);
-				continue;
-			}
+			mdb = mapregstr_db;
+		}
+		
+		if( i64db_exists(mdb, reference_uid(s,i)) ) {
+			ShowWarning("load_mapreg: duplicate! '%s' (%d) => '%s' skipping...\n",varname,index,value);
+			continue;
 		}
 		
 		m = ers_alloc(mapreg_ers, struct mapreg_save);
-		m->uid = (i<<24)|s;
+		m->uid = reference_uid(s,i);
 		m->save = false;
+		
 		if( varname[length-1] == '$' ) {
 			m->u.str = aStrdup(value);
-			idb_put(mapregstr_db, m->uid, m);
 		} else {
 			m->u.i = atoi(value);
-			idb_put(mapreg_db, m->uid, m);
+		}
+		
+		i64db_put(mdb, m->uid, m);
+		
+		if( i > 0 ) {
+			if( (m = i64db_get(mdb, reference_uid(s,SCRIPT_ARRAY_LEN_IDX))) ) {
+				if( varname[length-1] == '$' ) {
+					int len;
+					
+					if( (len = atoi(m->u.str)) < i ) {
+						char v[11];
+						
+						aFree(m->u.str);
+						snprintf(v,11,"%d",i+1);
+						m->u.str = aStrdup(v);
+					}
+				} else {
+					if( m->u.i < i )
+						m->u.i = i+1;
+				}
+			} else {
+				m = ers_alloc(mapreg_ers, struct mapreg_save);
+				m->uid = reference_uid(s,SCRIPT_ARRAY_LEN_IDX);
+				m->save = false;
+				
+				if( varname[length-1] == '$' ) {
+					char v[11];
+					
+					snprintf(v,11,"%d",i+1);
+					m->u.str = aStrdup(v);
+				} else {
+					m->u.i = i+1;
+				}
+				i64db_put(mdb, m->uid, m);
+			}
 		}
 	}
 	
@@ -202,8 +238,8 @@ static void script_save_mapreg(void) {
 		iter = db_iterator(mapreg_db);
 		for( m = dbi_first(iter); dbi_exists(iter); m = dbi_next(iter) ) {
 			if( m->save ) {
-				int num = (m->uid & 0x00ffffff);
-				int i   = (m->uid & 0xff000000) >> 24;
+				int num = (int32)(m->uid & 0xFFFFFFFF);
+				int i   = (int32)((m->uid >> 32) & 0xFFFFFFFF);
 				const char* name = script->get_str(num);
 
 				if( SQL_ERROR == SQL->Query(mmysql_handle, "UPDATE `%s` SET `value`='%d' WHERE `varname`='%s' AND `index`='%d' LIMIT 1", mapreg_table, m->u.i, name, i) )
@@ -219,8 +255,8 @@ static void script_save_mapreg(void) {
 		iter = db_iterator(mapregstr_db);
 		for( m = dbi_first(iter); dbi_exists(iter); m = dbi_next(iter) ) {
 			if( m->save ) {
-				int num = (m->uid & 0x00ffffff);
-				int i   = (m->uid & 0xff000000) >> 24;
+				int num = (int32)(m->uid & 0xFFFFFFFF);
+				int i   = (int32)((m->uid >> 32) & 0xFFFFFFFF);
 				const char* name = script->get_str(num);
 				char tmp_str2[2*255+1];
 
@@ -296,8 +332,8 @@ void mapreg_final(void) {
 }
 
 void mapreg_init(void) {
-	mapreg_db = idb_alloc(DB_OPT_BASE);
-	mapregstr_db = idb_alloc(DB_OPT_BASE);
+	mapreg_db = i64db_alloc(DB_OPT_BASE);
+	mapregstr_db = i64db_alloc(DB_OPT_BASE);
 	mapreg_ers = ers_new(sizeof(struct mapreg_save), "mapreg_sql.c::mapreg_ers", ERS_OPT_NONE);
 
 	script_load_mapreg();
